More updates

+++ src/main/java/montador/Memory.java
@@ -1,18 +1,18 @@
 package montador;
 
 public class Memory{
-    private short[] palavras = new short[4096];
+    private int[] palavras = new int[4096];
 
     public Memory(){
         for(int i = 0; i < 4096; i++){
-            this.palavras[i] = (short)0;
+            this.palavras[i] = (int)0;
         }  
     }
-    public void setPalavra(short data, int position){
+    public void setPalavra(int data, int position){
         this.palavras[position] = data;
     }
 
-    public short getPalavra(int position){
+    public int getPalavra(int position){
         return this.palavras[position];
     }
 }
